requirements.txt
Django==4.0.0
djangorestframework==3.12.4
django-cors-headers==3.7.0
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
python-dotenv==0.19.2
requests==2.26.0
pytz
gunicorn==20.1.0
